feat: support human-readable size format (512M, 1G) for max_multipart_memory
- Define MemorySize type with TextUnmarshaler supporting K/M/G suffixes
- Config now accepts "512M", "1G", "536870912" or plain integers
- Default changed to "32M"

+++ internal/server/server.go
@@ -35,7 +35,7 @@ func New(config Config) *Server {
 	// Set max multipart memory for file uploads (e.g., backup restore).
 	// Default 32 MB may be insufficient for large backup files.
 	if config.MaxMultipartMemory > 0 {
-		engine.MaxMultipartMemory = config.MaxMultipartMemory
+		engine.MaxMultipartMemory = int64(config.MaxMultipartMemory)
 	}
 
 	engine.Use(middleware.Recovery())
